requirements.txt
astor==0.8.1
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
docutils==0.16
gast==0.4.0
h5py==3.2.1
idna==2.10
importlib-metadata==3.7.3
keyring==23.0.0
kiwisolver==1.3.1
licsber==2.1.0
matplotlib==3.4.2
minio==7.0.3
numpy==1.20.3
opencv-python==4.5.2.54
packaging==20.9
paddlepaddle==2.1.0
Pillow==8.2.0
pkginfo==1.7.0
protobuf==3.17.3
pycryptodome==3.10.1
Pygments==2.8.1
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.60.0
twine==3.4.1
urllib3==1.26.5
webencodings==0.5.1
zipp==3.4.1
